✨ Add ignore for VirtualizedLists warning
There is some kind of conflict between the scrollview for the Query component and the flatlists called as children later on. After spending many hours trying to fix this I've decided that the "proper" solution (as defined on a github issue on rn from years ago) would require a lot of work and make the code less readable. So I'm not doing it and instead cheating for this specific warning.